// src/app/models/comprassumasdetalle.ts
import { Comprassumas } from "./comprassumas";

export class Comprassumasdetalle{

    idcomprasumadetalle?: string;
    monto?: number;
    descuento?: number;
    ice100?: number;
    icecreditofis?: number;
    descuentoice100?: number;
    idcomprasuma?: number;
    comprassumas?: Comprassumas;


}